Change Makefile paradigm to remove PyTorch cache image
This is done because of the previously mentioned bug where PyTorch kept being rebuilt with `main`/`master`.

The problem was not solved and this new scheme relies on build caches for speedy installation.

It still needs testing.

Also, I dislike the idea that it is now harder to create the build-complete image.

However, one could manually tag an intermediate image with build already complete if this is absolutely necessary.