Implement install speedup and change variable names
Implemented `apt` and `pip` speedups for the training image.

The train image will be built often, which makes the speedup worthwhile.

The speedup implementation is not very flexible.

It relies on CUDA base images from NVIDIA for link swapping.

Also, it only works for Korean images.

However, this is OK because the `train` image is made to be customized anyway.

Also, the TRAIN_IMAGE_NAME and TORCH_IMAGE_NAME variables were changed to TRAIN_NAME TORCH_NAME.

Finally, CuPy was removed, along with associated libraries, because most people do not use CuPy with PyTorch.